Attribute Macro pinned_init_macro::manual_init

#[manual_init]

Use this attribute on a struct with named fields to ensure safer pinned initialization of all the fields marked with #[init].

This attribute does several things, it:

#[pin_project]s your struct, structually pinning all fields with #[pin].

adds a constant type parameter of type bool with a default value of true. This constant type parameter indicates if your struct is in an initialized (and thus also pinned) state. A type alias {your-struct-name}Uninit is created to refer to the uninitialized variant more ergonomically, it should always be used instead of specifying the const parameter.

propagates that const parameter to all fields marked with #[init].

implements [TransmuteInto<{your-struct-name}>] for{your-struct-name}Uninit` and checks for layout equivalence between the two.

creates a custom type borrowing from your struct that is used as the OngoingInit type for the [BeginInit] trait.

implements [BeginInit] for your struct.

The only thing you need to implement is [PinnedInit].

Then you can safely, soundly and ergonomically initialize a value of such a struct behind an [OwnedUniquePtr<{your-struct-name}>]: TODO example